+++ src/supervisor/event_loop.rs
@@ -117,10 +117,18 @@ impl super::VpnSupervisor {
                         }
                         VpnCommand::Restart => {
                             self.handle_restart().await;
-                            if self.should_exit {
-                                info!("Exiting due to: {:?}", self.exit_reason);
+                            if self.exit_state.should_exit {
+                                info!("Exiting due to: {:?}", self.exit_state.reason);
                                 self.graceful_shutdown().await;
-                                std::process::exit(0);
+                                return;
+                            }
+                        }
+                        VpnCommand::Quit => {
+                            self.handle_quit().await;
+                            if self.exit_state.should_exit {
+                                info!("Exiting due to: {:?}", self.exit_state.reason);
+                                self.graceful_shutdown().await;
+                                return;
                             }
                         }
                     }
@@ -134,10 +142,10 @@ impl super::VpnSupervisor {
                 // Handle IPC commands
                 Some((cmd, response_tx)) = self.ipc_rx.recv() => {
                     self.handle_ipc_command(cmd, response_tx).await;
-                    if self.should_exit {
-                        info!("Exiting due to: {:?}", self.exit_reason);
+                    if self.exit_state.should_exit {
+                        info!("Exiting due to: {:?}", self.exit_state.reason);
                         self.graceful_shutdown().await;
-                        std::process::exit(0);
+                        return;
                     }
                 }